requirements.txt
fastapi==0.85.1
nanoid==2.0.0
pydantic==1.10.2
pymongo[tls,srv]==4.2.0
python-dotenv==0.21.0
uvicorn==0.11.3

